More logical code structure
