+++ pkg/roots/common.go
@@ -2,12 +2,18 @@ package roots
 
 import (
 	"errors"
+	"fmt"
 	"log"
+	"net/url"
 
 	"github.com/adem-wg/adem-proto/pkg/tokens"
+	"github.com/adem-wg/adem-proto/pkg/util"
 	"github.com/lestrrat-go/jwx/v3/jwk"
 )
 
+var ErrIssNoHostName = errors.New("issuer has no hostname")
+var ErrCertNotForIss = errors.New("certificate is not valid for issuer OI")
+var ErrCertNotForKey = errors.New("certificate is not valid for key")
 var ErrNoLogConfig = errors.New("no log claim")
 
 type CTQueryResult struct {
@@ -25,6 +31,30 @@ func VerifyBindingCerts(iss string, key jwk.Key, logs []*tokens.LogConfig) []CTQ
 		verified[i].Ok = VerifyBinding(verified[i], iss, key) == nil
 	}
 	return verified
+}
+
+// Verify that the rootKey is correctly bound to the issuer OI in the
+// certificate's subjects referenced by the CT query.
+func VerifyBinding(q CTQueryResult, issuer string, rootKey jwk.Key) error {
+	kid, err := tokens.CalcKID(rootKey)
+	if err != nil {
+		log.Print("could not calculate KID")
+		return err
+	}
+	issuerUrl, err := url.Parse(issuer)
+	if err != nil {
+		log.Print("could not parse issuer")
+		return err
+	} else if issuerUrl.Hostname() == "" {
+		return ErrIssNoHostName
+	}
+
+	if !util.Contains(q.subjects, issuerUrl.Hostname()) {
+		return ErrCertNotForIss
+	} else if !util.Contains(q.subjects, fmt.Sprintf("%s.adem-configuration.%s", kid, issuerUrl.Hostname())) {
+		return ErrCertNotForKey
+	}
+	return nil
 }
 
 // Verify that the hashes in the log configs are included in the respective CT
