+++ pkg/roots/v1.go
@@ -7,48 +7,16 @@ package roots
 import (
 	"context"
 	"errors"
-	"fmt"
 	"log"
-	"net/url"
 	"time"
 
-	"github.com/adem-wg/adem-proto/pkg/tokens"
-	"github.com/adem-wg/adem-proto/pkg/util"
 	"github.com/google/certificate-transparency-go/client"
 	"github.com/google/certificate-transparency-go/x509"
-	"github.com/lestrrat-go/jwx/v3/jwk"
 	"github.com/transparency-dev/merkle/proof"
 	"github.com/transparency-dev/merkle/rfc6962"
 )
 
-var ErrIssNoHostName = errors.New("issuer has no hostname")
-var ErrCertNotForIss = errors.New("certificate is not valid for issuer OI")
-var ErrCertNotForKey = errors.New("certificate is not valid for key")
 var ErrWrongEntryType = errors.New("do not recognize entry type")
-
-// Verify that the rootKey is correctly bound to the issuer OI in the
-// certificate's subjects referenced by the CT query.
-func VerifyBinding(q CTQueryResult, issuer string, rootKey jwk.Key) error {
-	kid, err := tokens.CalcKID(rootKey)
-	if err != nil {
-		log.Print("could not calculate KID")
-		return err
-	}
-	issuerUrl, err := url.Parse(issuer)
-	if err != nil {
-		log.Print("could not parse issuer")
-		return err
-	} else if issuerUrl.Hostname() == "" {
-		return ErrIssNoHostName
-	}
-
-	if !util.Contains(q.subjects, issuerUrl.Hostname()) {
-		return ErrCertNotForIss
-	} else if !util.Contains(q.subjects, fmt.Sprintf("%s.adem-configuration.%s", kid, issuerUrl.Hostname())) {
-		return ErrCertNotForKey
-	}
-	return nil
-}
 
 // Verify that the given certificate hash is included in the log identified by
 // the respective client.
